Preliminary commit with clientUtils and socketUtils.
