+++ index.js
@@ -1,7 +1,9 @@
 const cliClient = require('./src/cliClient')
+const wsClient = require('./src/websockClient')
 const server = require('./src/server')
 
 module.exports = {
   ...cliClient,
+  ...wsClient,
   ...server,
 }
